embedder/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
torchao
Pillow
python-multipart
